backend/requirements.txt
# Roboroça Backend - Dependencies
# FastAPI e Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image Processing
pillow>=10.2.0
numpy>=1.26.0
opencv-python-headless>=4.9.0

# Machine Learning (optional - install if needed)
# torch>=2.1.0
# ultralytics>=8.1.0

# Utilities
python-dateutil>=2.8.2
aiofiles>=23.2.0
httpx>=0.26.0

# PDF Generation
fpdf2>=2.7.0
